quant-engine/requirements.txt
fastapi
uvicorn
pydantic
scikit-learn
statsmodels
arch
yfinance
shap
numpy
pandas
hmmlearn
python-dotenv